requirements.txt



google-generativeai
mediapipe
numpy
python-dotenv